Bug 1485699: BrowserViewController cleanup: isolate ReaderView and WebKit delegates (#4167)
