+++ Client/Frontend/Browser/BrowserViewController/BrowserViewController+WebViewDelegates.swift
@@ -8,6 +8,90 @@ import Shared
 
 private let log = Logger.browserLogger
 
+/// List of schemes that are allowed to be opened in new tabs.
+private let schemesAllowedToBeOpenedAsPopups = ["http", "https", "javascript", "data", "about"]
+
+extension BrowserViewController: WKUIDelegate {
+    func webView(_ webView: WKWebView, createWebViewWith configuration: WKWebViewConfiguration, for navigationAction: WKNavigationAction, windowFeatures: WKWindowFeatures) -> WKWebView? {
+        guard let parentTab = tabManager[webView] else { return nil }
+
+        guard navigationAction.isAllowed, shouldRequestBeOpenedAsPopup(navigationAction.request) else {
+            print("Denying popup from request: \(navigationAction.request)")
+            return nil
+        }
+
+        if let currentTab = tabManager.selectedTab {
+            screenshotHelper.takeScreenshot(currentTab)
+        }
+
+        // If the page uses `window.open()` or `[target="_blank"]`, open the page in a new tab.
+        // IMPORTANT!!: WebKit will perform the `URLRequest` automatically!! Attempting to do
+        // the request here manually leads to incorrect results!!
+        let newTab = tabManager.addPopupForParentTab(parentTab, configuration: configuration)
+
+        return newTab.webView
+    }
+
+    fileprivate func shouldRequestBeOpenedAsPopup(_ request: URLRequest) -> Bool {
+        // Treat `window.open("")` the same as `window.open("about:blank")`.
+        if request.url?.absoluteString.isEmpty ?? false {
+            return true
+        }
+
+        if let scheme = request.url?.scheme?.lowercased(), schemesAllowedToBeOpenedAsPopups.contains(scheme) {
+            return true
+        }
+
+        return false
+    }
+
+    fileprivate func shouldDisplayJSAlertForWebView(_ webView: WKWebView) -> Bool {
+        // Only display a JS Alert if we are selected and there isn't anything being shown
+        return ((tabManager.selectedTab == nil ? false : tabManager.selectedTab!.webView == webView)) && (self.presentedViewController == nil)
+    }
+
+    func webView(_ webView: WKWebView, runJavaScriptAlertPanelWithMessage message: String, initiatedByFrame frame: WKFrameInfo, completionHandler: @escaping () -> Void) {
+        let messageAlert = MessageAlert(message: message, frame: frame, completionHandler: completionHandler)
+        if shouldDisplayJSAlertForWebView(webView) {
+            present(messageAlert.alertController(), animated: true, completion: nil)
+        } else if let promptingTab = tabManager[webView] {
+            promptingTab.queueJavascriptAlertPrompt(messageAlert)
+        } else {
+            // This should never happen since an alert needs to come from a web view but just in case call the handler
+            // since not calling it will result in a runtime exception.
+            completionHandler()
+        }
+    }
+
+    func webView(_ webView: WKWebView, runJavaScriptConfirmPanelWithMessage message: String, initiatedByFrame frame: WKFrameInfo, completionHandler: @escaping (Bool) -> Void) {
+        let confirmAlert = ConfirmPanelAlert(message: message, frame: frame, completionHandler: completionHandler)
+        if shouldDisplayJSAlertForWebView(webView) {
+            present(confirmAlert.alertController(), animated: true, completion: nil)
+        } else if let promptingTab = tabManager[webView] {
+            promptingTab.queueJavascriptAlertPrompt(confirmAlert)
+        } else {
+            completionHandler(false)
+        }
+    }
+
+    func webView(_ webView: WKWebView, runJavaScriptTextInputPanelWithPrompt prompt: String, defaultText: String?, initiatedByFrame frame: WKFrameInfo, completionHandler: @escaping (String?) -> Void) {
+        let textInputAlert = TextInputAlert(message: prompt, frame: frame, completionHandler: completionHandler, defaultText: defaultText)
+        if shouldDisplayJSAlertForWebView(webView) {
+            present(textInputAlert.alertController(), animated: true, completion: nil)
+        } else if let promptingTab = tabManager[webView] {
+            promptingTab.queueJavascriptAlertPrompt(textInputAlert)
+        } else {
+            completionHandler(nil)
+        }
+    }
+
+    func webViewDidClose(_ webView: WKWebView) {
+        if let tab = tabManager[webView] {
+            self.tabManager.removeTabAndUpdateSelectedIndex(tab)
+        }
+    }
+}
+
 extension WKNavigationAction {
     /// Allow local requests only if the request is privileged.
     var isAllowed: Bool {
